fix: drop admission hold across cleanup remove
Avoid holding admission.mu during cancellable CleanupWorktree so
MarkDegraded/BeginShutdown can cancel without deadlock, keep terminal
cleanup events gated under WithAllowClaim, and split admission cleanup
contract tests out of worktree_cleanup_test.go.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
