+++ internal/runtime/worktree_cleanup_test.go
@@ -349,469 +349,6 @@ func newWorktreeCleanupFixture(t *testing.T) worktreeCleanupFixture {
 	}
 }
 
-// Contract (#580 review): when admission is already closed, do not emit durable
-// cleanup events (including started) — only in-memory summary.
-func TestWorktreeCleanupPassOmitsEventsWhenAdmissionClosed(t *testing.T) {
-	t.Parallel()
-
-	fixture := newWorktreeCleanupFixture(t)
-	_ = fixture.seedWorktree(t, "wt_planned", "feature/planned", true)
-	// Gate start-event emission under WithAllowClaim so a closed admission
-	// cannot persist worktree.cleanup.started (or later cleanup events).
-	if err := fixture.runtime.admission.MarkDegraded("before start event"); err != nil {
-		t.Fatalf("MarkDegraded() error = %v", err)
-	}
-
-	git := &fakeWorktreeCleanupGit{
-		listed: map[string][]gitinfra.WorktreeListEntry{fixture.project.RepoPath: {}},
-		clean:  map[string]bool{},
-	}
-	summary := fixture.runtime.runWorktreeCleanupPass(context.Background(), fixture.repos, git, fixture.config)
-
-	if !strings.Contains(summary.LastError, "degraded") {
-		t.Fatalf("summary.LastError = %q, want admission degraded", summary.LastError)
-	}
-	if summary.Cleaned != 0 || summary.Skipped != 0 || summary.Failed != 0 {
-		t.Fatalf("summary = %#v, want no candidate mutations after admission closed", summary)
-	}
-	if len(git.cleanupCalls) != 0 {
-		t.Fatalf("cleanupCalls = %#v, want none while degraded", git.cleanupCalls)
-	}
-	events := fixture.events(t)
-	if containsWorktreeCleanupEvent(events, "worktree.cleanup.started") {
-		t.Fatalf("events = %#v, want no started event when admission closed at emission", events)
-	}
-	if containsWorktreeCleanupEvent(events, "worktree.cleanup.completed") {
-		t.Fatalf("events = %#v, want no completed event after admission closed", events)
-	}
-	if containsWorktreeCleanupEvent(events, "worktree.cleanup.skipped") || containsWorktreeCleanupEvent(events, "worktree.cleanup.cleaned") {
-		t.Fatalf("events = %#v, want no skip/cleaned events after admission closed", events)
-	}
-}
-
-// Contract (#580 / review): MarkDegraded mid-pass must refuse remaining cleanup
-// mutations. Concurrent MarkDegraded blocks on WithAllowClaim during the first
-// delete (cannot close admission mid-remove), then closes after the hold so the
-// second candidate never reaches CleanupWorktree.
-func TestWorktreeCleanupPassCancelsWhenAdmissionClosesMidPass(t *testing.T) {
-	t.Parallel()
-
-	fixture := newWorktreeCleanupFixture(t)
-	// Stagger UpdatedAt so plan order is stable (older first).
-	first := fixture.seedWorktreeAt(t, "wt_first", "feature/first", true, fixture.now.Add(-2*time.Hour))
-	second := fixture.seedWorktreeAt(t, "wt_second", "feature/second", true, fixture.now.Add(-time.Hour))
-	enteredFirst := make(chan struct{})
-	releaseFirst := make(chan struct{})
-	var cleaned []string
-	git := &fakeWorktreeCleanupGit{
-		listed: map[string][]gitinfra.WorktreeListEntry{fixture.project.RepoPath: {
-			{Path: first.WorktreePath, Branch: first.Branch},
-			{Path: second.WorktreePath, Branch: second.Branch},
-		}},
-		clean: map[string]bool{first.WorktreePath: true, second.WorktreePath: true},
-		onCleanup: func(input gitinfra.CleanupWorktreeInput) error {
-			cleaned = append(cleaned, input.WorktreePath)
-			if input.WorktreePath != first.WorktreePath {
-				t.Fatalf("CleanupWorktree called for %q after admission should be closed", input.WorktreePath)
-			}
-			close(enteredFirst)
-			<-releaseFirst
-			updated, err := fixture.repos.Worktrees.GetByID(context.Background(), first.ID)
-			if err != nil {
-				return err
-			}
-			nowISO := fixture.now.Format("2006-01-02T15:04:05.000Z")
-			updated.Status = "cleaned"
-			updated.CleanedAt = &nowISO
-			updated.UpdatedAt = nowISO
-			return fixture.repos.Worktrees.Upsert(context.Background(), *updated)
-		},
-	}
-
-	done := make(chan WorktreeCleanupStatus, 1)
-	go func() {
-		done <- fixture.runtime.runWorktreeCleanupPass(context.Background(), fixture.repos, git, fixture.config)
-	}()
-
-	select {
-	case <-enteredFirst:
-	case <-time.After(2 * time.Second):
-		t.Fatal("timed out waiting for first CleanupWorktree under admission hold")
-	}
-
-	degradeDone := make(chan error, 1)
-	go func() {
-		degradeDone <- fixture.runtime.MarkDegraded("mid-pass degrade")
-	}()
-	// MarkDegraded must block while WithAllowClaim holds admission during delete.
-	select {
-	case err := <-degradeDone:
-		t.Fatalf("MarkDegraded completed while CleanupWorktree held admission: %v", err)
-	case <-time.After(30 * time.Millisecond):
-	}
-
-	close(releaseFirst)
-
-	select {
-	case err := <-degradeDone:
-		if err != nil {
-			t.Fatalf("MarkDegraded() after hold release error = %v", err)
-		}
-	case <-time.After(2 * time.Second):
-		t.Fatal("MarkDegraded did not complete after CleanupWorktree released admission")
-	}
-
-	var summary WorktreeCleanupStatus
-	select {
-	case summary = <-done:
-	case <-time.After(2 * time.Second):
-		t.Fatal("cleanup pass did not return after mid-pass degrade")
-	}
-
-	if len(cleaned) != 1 {
-		t.Fatalf("cleaned paths = %#v, want exactly one candidate before admission closed", cleaned)
-	}
-	if summary.Cleaned != 1 {
-		t.Fatalf("summary.Cleaned = %d, want 1", summary.Cleaned)
-	}
-	if fixture.runtime.admission.State() != AdmissionDegraded {
-		t.Fatalf("admission.State() = %q, want degraded after mid-pass MarkDegraded", fixture.runtime.admission.State())
-	}
-	// LastError is set when the loop-level AllowClaim refuses; a later
-	// per-candidate WithAllowClaim skip may only increment Skipped.
-	if summary.LastError != "" && !strings.Contains(summary.LastError, "degraded") {
-		t.Fatalf("summary.LastError = %q, want empty or admission degraded", summary.LastError)
-	}
-	if summary.LastError == "" && summary.Skipped < 1 {
-		t.Fatalf("summary = %#v, want LastError degraded or at least one skipped candidate", summary)
-	}
-	stored, err := fixture.repos.Worktrees.GetByID(context.Background(), second.ID)
-	if err != nil {
-		t.Fatalf("Worktrees.GetByID(remaining) error = %v", err)
-	}
-	if stored == nil || stored.Status != "active" {
-		t.Fatalf("remaining worktree = %#v, want still active after mid-pass admission close", stored)
-	}
-}
-
-// Contract (#580 review): CleanupWorktree runs under WithAllowClaim so
-// MarkDegraded cannot close admission mid-remove. The in-flight delete may
-// finish under open admission (closer blocked on the same mutex); after the
-// hold releases, MarkDegraded proceeds and cancelWorkProducers runs too late
-// to retroactively forbid a remove that already completed under the hold.
-func TestWorktreeCleanupInFlightHoldsAdmissionAcrossCleanup(t *testing.T) {
-	t.Parallel()
-
-	fixture := newWorktreeCleanupFixture(t)
-	worktree := fixture.seedWorktree(t, "wt_inflight", "feature/inflight", true)
-
-	entered := make(chan struct{})
-	release := make(chan struct{})
-	git := &fakeWorktreeCleanupGit{
-		listed: map[string][]gitinfra.WorktreeListEntry{fixture.project.RepoPath: {
-			{Path: worktree.WorktreePath, Branch: worktree.Branch},
-		}},
-		clean: map[string]bool{worktree.WorktreePath: true},
-		onCleanup: func(input gitinfra.CleanupWorktreeInput) error {
-			close(entered)
-			<-release
-			updated, err := fixture.repos.Worktrees.GetByID(context.Background(), worktree.ID)
-			if err != nil {
-				return err
-			}
-			nowISO := fixture.now.Format("2006-01-02T15:04:05.000Z")
-			updated.Status = "cleaned"
-			updated.CleanedAt = &nowISO
-			updated.UpdatedAt = nowISO
-			return fixture.repos.Worktrees.Upsert(context.Background(), *updated)
-		},
-	}
-
-	done := make(chan WorktreeCleanupStatus, 1)
-	go func() {
-		done <- fixture.runtime.runWorktreeCleanupPass(context.Background(), fixture.repos, git, fixture.config)
-	}()
-
-	select {
-	case <-entered:
-	case <-time.After(2 * time.Second):
-		t.Fatal("timed out waiting for in-flight CleanupWorktree under admission hold")
-	}
-
-	degradeDone := make(chan error, 1)
-	go func() {
-		degradeDone <- fixture.runtime.MarkDegraded("mid cleanup pass")
-	}()
-	// Do not call admission.State() here: it takes the same mutex WithAllowClaim
-	// holds for CleanupWorktree and would deadlock the test.
-	select {
-	case err := <-degradeDone:
-		t.Fatalf("MarkDegraded completed while CleanupWorktree held admission: %v", err)
-	case <-time.After(30 * time.Millisecond):
-	}
-
-	close(release)
-
-	select {
-	case err := <-degradeDone:
-		if err != nil {
-			t.Fatalf("MarkDegraded() after hold release error = %v", err)
-		}
-	case <-time.After(2 * time.Second):
-		t.Fatal("MarkDegraded did not complete after CleanupWorktree released admission")
-	}
-
-	var summary WorktreeCleanupStatus
-	select {
-	case summary = <-done:
-	case <-time.After(2 * time.Second):
-		t.Fatal("cleanup pass did not return after held cleanup finished")
-	}
-
-	// Delete completed under open admission (closer was blocked); cleaned=1 is correct.
-	if summary.Cleaned != 1 {
-		t.Fatalf("summary.Cleaned = %d, want 1 (delete finished under held open admission)", summary.Cleaned)
-	}
-	if state := fixture.runtime.admission.State(); state != AdmissionDegraded {
-		t.Fatalf("admission.State() after pass = %q, want degraded", state)
-	}
-	stored, err := fixture.repos.Worktrees.GetByID(context.Background(), worktree.ID)
-	if err != nil {
-		t.Fatalf("Worktrees.GetByID() error = %v", err)
-	}
-	if stored == nil || stored.Status != "cleaned" {
-		t.Fatalf("stored worktree = %#v, want cleaned after held cleanup completed", stored)
-	}
-	events := fixture.events(t)
-	if !containsWorktreeCleanupEvent(events, "worktree.cleanup.cleaned") {
-		t.Fatalf("events = %#v, want cleaned event under held admission", events)
-	}
-}
-
-// Contract: destructive CleanupWorktree is refused when admission closes after
-// eligibility checks but before the delete.
-func TestCleanWorktreeCandidateRefusesWhenAdmissionClosed(t *testing.T) {
-	t.Parallel()
-
-	fixture := newWorktreeCleanupFixture(t)
-	worktree := fixture.seedWorktree(t, "wt_closed", "feature/closed", true)
-	if err := fixture.runtime.admission.MarkDegraded("before delete"); err != nil {
-		t.Fatalf("MarkDegraded() error = %v", err)
-	}
-	git := &fakeWorktreeCleanupGit{
-		listed: map[string][]gitinfra.WorktreeListEntry{fixture.project.RepoPath: {{Path: worktree.WorktreePath, Branch: worktree.Branch}}},
-		clean:  map[string]bool{worktree.WorktreePath: true},
-	}
-
-	result := fixture.runtime.cleanWorktreeCandidate(context.Background(), fixture.repos, git, fixture.config, fixture.project, worktree, fixture.root, "clean")
-	if result.status != "skipped" || !strings.Contains(result.message, "degraded") {
-		t.Fatalf("result = %#v, want skipped admission degraded", result)
-	}
-	if len(git.cleanupCalls) != 0 {
-		t.Fatalf("cleanupCalls = %#v, want none while degraded", git.cleanupCalls)
-	}
-}
-
-// Contract (#592 review): Plan failure terminal events (failed/completed) are
-// held under WithAllowClaim so a closed admission cannot append them after
-// MarkDegraded cancels the cleanup context mid-Plan.
-func TestWorktreeCleanupPassGatesPlanFailureTerminalEvents(t *testing.T) {
-	t.Parallel()
-
-	fixture := newWorktreeCleanupFixture(t)
-	// Force Plan to fail after the start-event gate without canceling ctx, so
-	// event appends are not masked by context.Canceled on ExecContext.
-	fixture.repos.Worktrees = nil
-	git := &fakeWorktreeCleanupGit{}
-
-	// Open admission: plan failure still records terminal events under the hold.
-	summary := fixture.runtime.runWorktreeCleanupPass(context.Background(), fixture.repos, git, fixture.config)
-	if summary.LastStatus != "failed" || summary.Failed != 1 {
-		t.Fatalf("summary = %#v, want failed status with Failed=1", summary)
-	}
-	if summary.LastError == "" {
-		t.Fatal("summary.LastError empty, want plan failure message")
-	}
-	events := fixture.events(t)
-	if !containsWorktreeCleanupEvent(events, "worktree.cleanup.started") {
-		t.Fatalf("events = %#v, want started before plan failure", events)
-	}
-	if !containsWorktreeCleanupEvent(events, "worktree.cleanup.failed") || !containsWorktreeCleanupEvent(events, "worktree.cleanup.completed") {
-		t.Fatalf("events = %#v, want failed+completed while admission open", events)
-	}
-
-	// Closed admission before the pass: start gate refuses; no durable events
-	// from a second attempt (prior events remain from the open-admission pass).
-	if err := fixture.runtime.admission.MarkDegraded("before plan failure pass"); err != nil {
-		t.Fatalf("MarkDegraded() error = %v", err)
-	}
-	before := len(fixture.events(t))
-	summary2 := fixture.runtime.runWorktreeCleanupPass(context.Background(), fixture.repos, git, fixture.config)
-	if !strings.Contains(summary2.LastError, "degraded") {
-		t.Fatalf("summary2.LastError = %q, want admission degraded", summary2.LastError)
-	}
-	after := fixture.events(t)
-	if len(after) != before {
-		t.Fatalf("events grew from %d to %d after closed-admission plan path; events = %#v", before, len(after), after)
-	}
-}
-
-// Contract (#592 review): when MarkDegraded cancels the pass context mid-pass,
-// the ctx.Err() break must not append worktree.cleanup.completed after admission
-// has already closed (cancelWorkProducers runs after the degraded transition).
-func TestWorktreeCleanupPassOmitsCompletedWhenContextCanceledAfterDegrade(t *testing.T) {
-	t.Parallel()
-
-	fixture := newWorktreeCleanupFixture(t)
-	first := fixture.seedWorktreeAt(t, "wt_ctx_first", "feature/ctx-first", true, fixture.now.Add(-2*time.Hour))
-	second := fixture.seedWorktreeAt(t, "wt_ctx_second", "feature/ctx-second", true, fixture.now.Add(-time.Hour))
-
-	ctx, cancel := context.WithCancel(context.Background())
-	// Production MarkDegraded cancels worktreeCleanupCancel; wire it so the
-	// in-flight pass observes ctx.Err() after admission closes.
-	fixture.runtime.worktreeCleanupCancel = cancel
-
-	enteredFirst := make(chan struct{})
-	releaseFirst := make(chan struct{})
-	git := &fakeWorktreeCleanupGit{
-		listed: map[string][]gitinfra.WorktreeListEntry{fixture.project.RepoPath: {
-			{Path: first.WorktreePath, Branch: first.Branch},
-			{Path: second.WorktreePath, Branch: second.Branch},
-		}},
-		clean: map[string]bool{first.WorktreePath: true, second.WorktreePath: true},
-		onCleanup: func(input gitinfra.CleanupWorktreeInput) error {
-			if input.WorktreePath != first.WorktreePath {
-				t.Fatalf("CleanupWorktree called for %q; second candidate must not mutate after cancel", input.WorktreePath)
-			}
-			close(enteredFirst)
-			<-releaseFirst
-			updated, err := fixture.repos.Worktrees.GetByID(context.Background(), first.ID)
-			if err != nil {
-				return err
-			}
-			nowISO := fixture.now.Format("2006-01-02T15:04:05.000Z")
-			updated.Status = "cleaned"
-			updated.CleanedAt = &nowISO
-			updated.UpdatedAt = nowISO
-			return fixture.repos.Worktrees.Upsert(context.Background(), *updated)
-		},
-	}
-
-	done := make(chan WorktreeCleanupStatus, 1)
-	go func() {
-		done <- fixture.runtime.runWorktreeCleanupPass(ctx, fixture.repos, git, fixture.config)
-	}()
-
-	select {
-	case <-enteredFirst:
-	case <-time.After(2 * time.Second):
-		t.Fatal("timed out waiting for first CleanupWorktree under admission hold")
-	}
-
-	degradeDone := make(chan error, 1)
-	go func() {
-		degradeDone <- fixture.runtime.MarkDegraded("cancel cleanup mid-pass")
-	}()
-	select {
-	case err := <-degradeDone:
-		t.Fatalf("MarkDegraded completed while CleanupWorktree held admission: %v", err)
-	case <-time.After(30 * time.Millisecond):
-	}
-
-	close(releaseFirst)
-
-	select {
-	case err := <-degradeDone:
-		if err != nil {
-			t.Fatalf("MarkDegraded() after hold release error = %v", err)
-		}
-	case <-time.After(2 * time.Second):
-		t.Fatal("MarkDegraded did not complete after CleanupWorktree released admission")
-	}
-
-	var summary WorktreeCleanupStatus
-	select {
-	case summary = <-done:
-	case <-time.After(2 * time.Second):
-		t.Fatal("cleanup pass did not return after mid-pass cancel")
-	}
-
-	if summary.Cleaned != 1 {
-		t.Fatalf("summary.Cleaned = %d, want 1 (first candidate finished under hold)", summary.Cleaned)
-	}
-	if fixture.runtime.admission.State() != AdmissionDegraded {
-		t.Fatalf("admission.State() = %q, want degraded", fixture.runtime.admission.State())
-	}
-	events := fixture.events(t)
-	if !containsWorktreeCleanupEvent(events, "worktree.cleanup.started") {
-		t.Fatalf("events = %#v, want started", events)
-	}
-	if !containsWorktreeCleanupEvent(events, "worktree.cleanup.cleaned") {
-		t.Fatalf("events = %#v, want cleaned under held open admission", events)
-	}
-	// Terminal completed must not append after admission closed via cancel path.
-	if containsWorktreeCleanupEvent(events, "worktree.cleanup.completed") {
-		t.Fatalf("events = %#v, want no completed after canceled pass under closed admission", events)
-	}
-	if containsWorktreeCleanupEvent(events, "worktree.cleanup.failed") {
-		t.Fatalf("events = %#v, want no pass-level failed after cancel under closed admission", events)
-	}
-}
-
-// Contract (#580 review): plan-skip event append is under WithAllowClaim so a
-// closed admission cannot persist worktree.cleanup.skipped after close.
-func TestWorktreeCleanupPlanSkipHoldsAdmission(t *testing.T) {
-	t.Parallel()
-
-	fixture := newWorktreeCleanupFixture(t)
-	worktree := fixture.seedWorktree(t, "wt_plan_skip", "feature/plan-skip", true)
-	if err := fixture.runtime.admission.MarkDegraded("before plan skip"); err != nil {
-		t.Fatalf("MarkDegraded() error = %v", err)
-	}
-
-	err := fixture.runtime.recordWorktreeCleanupPlanSkip(context.Background(), fixture.repos, worktree, "below_min_age")
-	if !errors.Is(err, ErrAdmissionDegraded) {
-		t.Fatalf("recordWorktreeCleanupPlanSkip() = %v, want ErrAdmissionDegraded", err)
-	}
-	events := fixture.events(t)
-	if containsWorktreeCleanupEvent(events, "worktree.cleanup.skipped") {
-		t.Fatalf("events = %#v, want no skip event when admission closed at write boundary", events)
-	}
-}
-
-// Contract (#580 review): candidate skip/failure record helpers hold admission
-// across worktrees touch + event append so degradation after eligibility checks
-// cannot commit durable cleanup mutations after close.
-func TestWorktreeCleanupRecordHelpersHoldAdmission(t *testing.T) {
-	t.Parallel()
-
-	fixture := newWorktreeCleanupFixture(t)
-	worktree := fixture.seedWorktree(t, "wt_record_gate", "feature/record-gate", true)
-	if err := fixture.runtime.admission.MarkDegraded("before record write"); err != nil {
-		t.Fatalf("MarkDegraded() error = %v", err)
-	}
-
-	skip := fixture.runtime.recordWorktreeCleanupSkip(context.Background(), fixture.repos, worktree, "dirty_git_status")
-	if skip.status != "skipped" || !strings.Contains(skip.message, "degraded") {
-		t.Fatalf("recordWorktreeCleanupSkip() = %#v, want skipped admission degraded", skip)
-	}
-	failure := fixture.runtime.recordWorktreeCleanupFailure(context.Background(), fixture.repos, worktree, errors.New("git list failed"))
-	if failure.status != "skipped" || !strings.Contains(failure.message, "degraded") {
-		t.Fatalf("recordWorktreeCleanupFailure() = %#v, want skipped admission degraded", failure)
-	}
-
-	stored, err := fixture.repos.Worktrees.GetByID(context.Background(), worktree.ID)
-	if err != nil {
-		t.Fatalf("Worktrees.GetByID() error = %v", err)
-	}
-	if stored == nil || stored.UpdatedAt != worktree.UpdatedAt {
-		t.Fatalf("stored worktree = %#v, want UpdatedAt unchanged after admission closed (no touch)", stored)
-	}
-	events := fixture.events(t)
-	if containsWorktreeCleanupEvent(events, "worktree.cleanup.skipped") || containsWorktreeCleanupEvent(events, "worktree.cleanup.failed") {
-		t.Fatalf("events = %#v, want no skip/failed events when admission closed at write boundary", events)
-	}
-}
 
 func (f worktreeCleanupFixture) seedWorktree(t *testing.T, id, branch string, createDir bool) storage.WorktreeRecord {
 	t.Helper()
